+++ web/Changelog.md
@@ -5,3 +5,46 @@
 Created separation between example pages, and actual pages/components.
 
 Trimmed some unneccesary features within the application.
+
+Added Login Page, and routed to main dashboard.
+
+Added Material UI, JSS depedency
+
+## Version 1.0.0 (27 March 2020)
+
+Added Firebase dependency.
+
+Updated all depdencies and went from Webpack 3 to 4. Webpack Dev Server from 2 to 3.
+
+## Version 1.0.0 (30 March 2020)
+
+Added Swagger API
+
+Added Chartjs into Dashboard
+
+Fixed bug that login button and sign in do not work.
+
+Started Firebase Database Project (Real-Time Database)
+
+## Version 1.0.0 (31 March 2020)
+
+Fixed Webpack 4 issue, to explicity set mode to development.
+
+Added Sign Up Firebase Authentication.
+
+Added Pills + Contact Doctors + Schedule Page.
+
+Added new pages above to app side bar navigation.
+
+Added User Profile into Firebase Real-Time Database.
+
+Added user centric UI (Added fname and lname to be displayed).
+
+Hide developer options from users.
+
+Added Checkbox for Cargiver in Sign Up.
+
+Added Check User Analysis Page + Set Users Schedule Page + Message Users Page
+
+## Version 1.0.0 (1 April 2020)
+